Fix potential use-after-free and MT-unsafe operation
The save-callback APIs are not thread-safe and should be called only
from the UI thread.
Passing the source data pointer to the UI thread risks a use-after-free
of the pointer if the source is released before the UI thread processed
the pointer.
This commit introduces global variables to store information retrieved
from the global configuration so that the UI thread never accesses the
source data.